% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcdeg.R
\name{calcdeg}
\alias{calcdeg}
\title{A function to perform RNA seq DEG analysis with up to four additional covariates and results annotation}
\usage{
calcdeg(dgeObjx, outcome, cov1 = NULL, cov2 = NULL, cov3 = NULL, cov4 = NULL)
}
\arguments{
\item{dgeObjx}{DGEList from previous analysis}

\item{outcome}{outcome from sampleinfo}

\item{cov1}{covariate in sampleinfo default is NULL}

\item{cov2}{covariate in sampleinfo  default is NULL}

\item{cov3}{covariate in sampleinfo  default is NULL}

\item{cov4}{covariate in sampleinfo  default is NULL}
}
\value{
annotated results
}
\description{
A function to perform RNA seq DEG analysis with up to four additional covariates and results annotation
}
