% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myrnaseq.R
\name{myrnaseq}
\alias{myrnaseq}
\title{A function to perform RNA seq DEG analysis}
\usage{
myrnaseq(countdata, sampleinfo, samplenum, groupnum)
}
\arguments{
\item{countdata}{obtained raw counts}

\item{sampleinfo}{file with mandatory samples(names) and group(what to compare) columns}

\item{samplenum}{number of all samples}

\item{groupnum}{number of samples in smallest group}
}
\value{
deg dataframe your_object$degs and dgeObj your_object$dgeObj
}
\description{
A function to perform RNA seq DEG analysis
}
