% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano.R
\name{volcano}
\alias{volcano}
\title{A function to draw volcano plots}
\usage{
volcano(myresults, Q = FALSE, top = 5, xl = NULL, xu = NULL)
}
\arguments{
\item{myresults}{annotated results from calcdeg}

\item{Q}{if TRUE take Q values, default is FALSE}

\item{top}{nu,ber of top genes to annotate, default is 5}

\item{xl}{lower x axis value, must be negative, default is NULL}

\item{xu}{upper x axis value, must be positive, default is NULL}
}
\value{
volcanoplot
}
\description{
A function to draw volcano plots
}
